+++ src/Vector/Vector.ts
@@ -1,6 +1,6 @@
-import { None, Option, Some } from "../Option";
 import { Panic } from "../panic";
 import { range } from "../utils";
+import { None, Option, Some } from "../Option";
 
 type Enumerate<T> = [index: number, value: T];
