+++ src/utils/unwrap-failed.ts
@@ -1,0 +1,4 @@
+/** @ignore */
+export function unwrapFailed<T>(msg: string, data: T): never {
+	throw new Error(`${msg}: ${JSON.stringify(data)}`);
+}
